+++ tests/testData/integration/arguments/long.m
@@ -2,10 +2,10 @@
 #import <stdio.h>
 
 @implementation A
-+ (void) foo: (long) arg {
++ (void) foo: (long long) arg {
     if (arg == 123456789123456789L)
         printf("OK");
     else
-        printf("Fail %ld", arg);
+        printf("Fail %lld", arg);
 }
 @end
