+++ tests/testData/integration/returnType/long.h
@@ -1,6 +1,6 @@
 #import <Foundation/NSObject.h>
 
 @interface A : NSObject
-+ (long) foo;
++ (long long) foo;
 + (long long) bar;
 @end
